实现mod及其依赖自动下载
1. 移除不必要的依赖
2. 添加已存在文件检查及其配置项
3. 实现搜索File
4. 添加枚举RelationType
5. SearchModsCondition中的val改为var
6. File的长度和下载次数改为long类型
7. Pagination添加了检测是否有下一页的方法
8. 实现了search和download命令

+++ src/main/kotlin/com/undsf/mcmodmgr/curseforge/responses/File.kt
@@ -38,10 +38,10 @@ data class File(
         var fileDate: LocalDateTime,
 
         @JsonProperty
-        var fileLength: Int,
+        var fileLength: Long,
 
         @JsonProperty
-        var downloadCount: Int,
+        var downloadCount: Long,
 
         @JsonProperty
         var downloadUrl: String?,
